requirements.txt
python-whois
pandas
matplotlib
numpy
scipy
scikit-learn
streamlit